Fixed onMessage
